+++ app/src/main/java/com/android/voyageur/ui/formFields/DatePicker.kt
@@ -1,0 +1,34 @@
+package com.android.voyageur.ui.formFields
+
+import androidx.compose.material3.DatePicker
+import androidx.compose.material3.DatePickerDialog
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.rememberDatePickerState
+import androidx.compose.runtime.Composable
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun DatePickerModal(
+    onDateSelected: (Long?) -> Unit,
+    onDismiss: () -> Unit,
+    selectedDate: Long = System.currentTimeMillis()
+) {
+  val datePickerState = rememberDatePickerState(initialSelectedDateMillis = selectedDate)
+
+  DatePickerDialog(
+      onDismissRequest = onDismiss,
+      confirmButton = {
+        TextButton(
+            onClick = {
+              onDateSelected(datePickerState.selectedDateMillis)
+              onDismiss()
+            }) {
+              Text("OK")
+            }
+      },
+      dismissButton = { TextButton(onClick = onDismiss) { Text("Cancel") } }) {
+        DatePicker(state = datePickerState)
+      }
+}
