+++ app/src/main/java/com/android/voyageur/ui/overview/AddTrip.kt
@@ -31,7 +31,7 @@ import com.android.voyageur.model.location.Location
 import com.android.voyageur.model.trip.Trip
 import com.android.voyageur.model.trip.TripType
 import com.android.voyageur.model.trip.TripsViewModel
-import com.android.voyageur.ui.DatePickerModal
+import com.android.voyageur.ui.formFields.DatePickerModal
 import com.android.voyageur.ui.navigation.NavigationActions
 import com.google.firebase.Firebase
 import com.google.firebase.Timestamp
